feat: Add getHints API and handleGenerate function
This commit introduces a new API endpoint, , which is used to generate hints based on user input. It also adds a new function, , which handles the generation of hints when the user clicks the Generate button. This function also manages a loading state to provide feedback to the user while the hints are being generated.